Split big C# code files (#682)
* Split HomePage.cs

* Split MainPage.cs

* [Sec] Define system executables path using SpecialFolder

* Backport changes from c88f7da13c73ea8b055ec239a4bf3af17b1c14ee


* Remove commented codes

* Adjust code comments for GameBoost_Invoke method

---------
